+++ backend/bang-ggood/src/main/java/com/bang_ggood/category/controller/CategoryController.java
@@ -1,7 +1,7 @@
 package com.bang_ggood.category.controller;
 
-import com.bang_ggood.category.dto.CategoriesReadResponse;
-import com.bang_ggood.category.dto.CategoryPriorityCreateRequest;
+import com.bang_ggood.category.dto.response.CategoriesReadResponse;
+import com.bang_ggood.category.dto.request.CategoryPriorityCreateRequest;
 import com.bang_ggood.category.service.CategoryService;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.GetMapping;
