+++ backend/bang-ggood/src/main/java/com/bang_ggood/category/dto/request/CategoryPriorityCreateRequest.java
@@ -1,4 +1,4 @@
-package com.bang_ggood.category.dto;
+package com.bang_ggood.category.dto.request;
 
 import java.util.List;
