+++ backend/bang-ggood/src/main/java/com/bang_ggood/category/dto/response/WrittenCategoryQuestionsResponse.java
@@ -1,6 +1,6 @@
-package com.bang_ggood.category.dto;
+package com.bang_ggood.category.dto.response;
 
-import com.bang_ggood.checklist.dto.WrittenQuestionResponse;
+import com.bang_ggood.checklist.dto.response.WrittenQuestionResponse;
 import java.util.List;
 
 public record WrittenCategoryQuestionsResponse(Integer categoryId, String categoryName, List<WrittenQuestionResponse> questions) {
