--- backend/bang-ggood/src/main/java/com/bang_ggood/checklist/dto/UserChecklistsPreviewResponse.java
@@ -1,6 +1,0 @@
-package com.bang_ggood.checklist.dto;
-
-import java.util.List;
-
-public record UserChecklistsPreviewResponse(List<UserChecklistPreviewResponse> checklists) {
-}
