+++ backend/bang-ggood/src/main/java/com/bang_ggood/checklist/dto/request/ChecklistCreateRequest.java
@@ -1,7 +1,7 @@
-package com.bang_ggood.checklist.dto;
+package com.bang_ggood.checklist.dto.request;
 
 import com.bang_ggood.room.domain.Room;
-import com.bang_ggood.room.dto.RoomCreateRequest;
+import com.bang_ggood.room.dto.request.RoomCreateRequest;
 import jakarta.validation.Valid;
 import java.util.List;
