+++ backend/bang-ggood/src/main/java/com/bang_ggood/checklist/dto/request/ChecklistInfo.java
@@ -1,4 +1,4 @@
-package com.bang_ggood.checklist.dto;
+package com.bang_ggood.checklist.dto.request;
 
 public record ChecklistInfo(Integer deposit, Integer rent, Integer contractTerm, String realEstate) {
 }
