+++ backend/bang-ggood/src/main/java/com/bang_ggood/checklist/dto/response/ChecklistsComparisonReadResponse.java
@@ -1,5 +1,6 @@
-package com.bang_ggood.checklist.dto;
+package com.bang_ggood.checklist.dto.response;
 
+import com.bang_ggood.checklist.dto.response.ChecklistComparisonReadResponse;
 import java.util.List;
 
 public record ChecklistsComparisonReadResponse(List<ChecklistComparisonReadResponse> checklists) {
