+++ backend/bang-ggood/src/main/java/com/bang_ggood/checklist/dto/response/UserChecklistsPreviewResponse.java
@@ -1,0 +1,7 @@
+package com.bang_ggood.checklist.dto.response;
+
+import com.bang_ggood.checklist.dto.response.UserChecklistPreviewResponse;
+import java.util.List;
+
+public record UserChecklistsPreviewResponse(List<UserChecklistPreviewResponse> checklists) {
+}
